#101: renamed components for clarity
