+++ src/webui/Routes.tsx
@@ -2,9 +2,11 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 
+import { useAppContext } from './contexts/AppContext';
+
 import Home from './features/Home';
-import GetIn from './features/GetIn';
 import Login from './features/Login';
+import Identify from './features/Identify';
 import Timeline from './features/Timeline';
 import Explore from './features/Explore';
 import Notifications from './features/Notifications';
@@ -14,15 +16,26 @@ import NotFound from './features/NotFound';
 
 export default function Component()
 {
+    const { identity } = useAppContext();
+
+    const protect = (node: React.ReactNode) =>
+    {
+        return identity === undefined ? <Login /> : node;
+    };
+
     return <Routes>
+
         <Route path="/" element={<Home />} />
-        <Route path="/getin" element={<GetIn />} />
         <Route path="/login" element={<Login />} />
-        <Route path="/timeline" element={<Timeline />} />
-        <Route path="/explore" element={<Explore />} />
-        <Route path="/notifications" element={<Notifications />} />
-        <Route path="/create" element={<CreateComic />} />
-        <Route path="/profile/:nickname" element={<Profile />} />
+        <Route path="/identify" element={<Identify />} />
+
+        <Route path="/timeline" element={protect(<Timeline />)} />
+        <Route path="/explore" element={protect(<Explore />)} />
+        <Route path="/notifications" element={protect(<Notifications />)} />
+        <Route path="/create" element={protect(<CreateComic />)} />
+        <Route path="/profile/:nickname" element={protect(<Profile />)} />
+
         <Route path="*" element={<NotFound />} />
+
     </Routes>;
 }
