+++ src/webui/features/Login.tsx
@@ -1,27 +1,16 @@
 
 import React, { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
 
-import johnDoe from '../../domain/authentication/johnDoe';
-import getMe from '../../domain/creator/getMe';
-
-import { useAppContext } from '../contexts/AppContext';
+import getLoginUrl from '../../domain/authentication/getLoginUrl';
 
 export default function Feature()
 {
-    const navigate = useNavigate();
-    const context = useAppContext();
-
-    const logMeIn = async () =>
+    const redirect = async () =>
     {
-        const me = await getMe(johnDoe);
-
-        context.setIdentity(me);
-
-        navigate('/timeline');
+        location.href = await getLoginUrl();
     };
 
-    useEffect(() => { logMeIn(); }, []);
+    useEffect(() => { redirect(); }, []);
 
-    return <>Logging in...</>;
+    return <>Redirecting...</>;
 }
